don't generate normal module when deriver is applied to a type that is not manifest
when the deriver is applied to a type like

type t = M.t = .... etc ...

it will generate a "normal" module with a copy of the type,
carrying-over the type-equation ("= M.t") so the type there is equal
to the original.  This is useful when the type was produced by
pa_ppx.import ("%import").  But when the type is directly written, it
might not be manifest, in which case we should not generate the
"normal" module (actually we *cannot* generate it).

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch master
# Your branch is up to date with 'origin/master'.
#
# Changes to be committed:
#	modified:   pa_hashcons/pa_deriving_hashcons.ml
#	modified:   tests/lamh.ml
#	modified:   tests/test_hashcons.ml
#
# Untracked files:
#	tests/include_ml
#